large addition see comments
added the backbone for adding tweets into the system. TweetsIO and
Tweets, one is a object class the other is for adding and retrieving
tweets.

worked on style sheet some more. broke some fixed others, planning on
working on style more after more functionality is added

home page should work once the respective servlet is added.

I plan on working on this some more tonight but not until late. need a
break or im going to break this stylesheet.